Updated to work with the latest BoxVR as of 2021-03-20
After decompiling box vr and going through the playlist and track beatmap generation, I've extracted the code that does it and adapted to the UI of the playlist manager.
Had to remove some functionality as didn't have time yet to properly implement it (importing of playlists, drag and drop)
This still needs more work as there are unused methods, properties and the code is a little too obscure, but this is a minimal viable product, so people can start using it.